feat(admin): mejorar la gestión de productos y la interfaz de administración
- Se actualizó el controlador AdicionalController para manejar las vistas de administración de adicionales.
- Se implementaron nuevas funcionalidades en AdminController para estadísticas y gestión de adicionales.
- Se mejoró la estructura de los modelos Cliente y Producto, añadiendo nuevos campos y métodos.
- Se eliminaron archivos innecesarios y se optimizó el código en varios controladores y servicios.
- Se añadieron estilos CSS específicos para la interfaz de administración y se mejoró la experiencia de usuario en el menú.
- Se actualizaron los archivos de configuración para un mejor rendimiento y depuración.

+++ burger-club/burgur/src/main/java/restaurante/example/burgur/Repository/ProductoRepository.java
@@ -1,3 +1,4 @@
+//burger-club/burgur/src/main/java/restaurante/example/burgur/Repository/ProductoRepository.java
 package restaurante.example.burgur.Repository;
 
 import java.util.List;
@@ -63,11 +64,5 @@ public interface ProductoRepository extends JpaRepository<Producto, Long> {
     // ==========================================
 
     // verificar si existe con el mismo nombre pero distinto id
-    boolean existsByNombreIgnoreCaseAndIdNot(String nombre, Integer excludeId);
-
-
-
-
-
-
+    boolean existsByNombreIgnoreCaseAndIdNot(String nombre, Long excludeId);
 }
